+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ### 1.0.1
 * added `fit()` method for `ThumbCreator` and `ThumbHelper` classes;
 * the `save()` method can save thumbnails with a different format;
-* the `save()` method has the `quality` option (default: 90).
+* the `save()` method has the `quality` option (default: 90);
+* by default, the `ThumbHelper` creates jpg thumbnails.
 
 ### 1.0.0
 * first release.